move types and helper to common
Update transition chaire-lib to get CliUser
This allows the helpers and types to be used both in backend and frontend, and also use the interview types for the questionnaire config instead of creating separated config types
