+++ example/demo_survey/src/config/widgets.config.ts
@@ -1,4 +1,4 @@
-import { WidgetConfig } from "evolution-frontend/lib/services/widgets";
+import { WidgetConfig } from "evolution-common/lib/services/widgets";
 
 // Types for the various custom responses. Note that for complete type check,
 // ALL responses should be typed and no `[key: string]: type` fields otherwise
