+++ packages/evolution-frontend/src/components/inputs/InputCheckbox.tsx
@@ -10,10 +10,10 @@ import { shuffle } from 'chaire-lib-common/lib/utils/RandomUtils';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { _chunkify } from 'chaire-lib-common/lib/utils/LodashExtensions';
-import * as surveyHelper from '../../utils/helpers';
-import { InputCheckboxType, ChoiceType } from '../../services/widgets';
+import * as surveyHelper from 'evolution-common/lib/utils/helpers';
+import { InputCheckboxType, ChoiceType } from 'evolution-common/lib/services/widgets';
 import { UserInterviewAttributes } from 'evolution-common/lib/services/interviews/interview';
-import { FrontendUser } from 'chaire-lib-frontend/lib/services/auth/user';
+import { CliUser } from 'chaire-lib-common/lib/services/user/userType';
 
 export interface InputCheckboxProps<CustomSurvey, CustomHousehold, CustomHome, CustomPerson> {
     id: string;
@@ -24,7 +24,7 @@ export interface InputCheckboxProps<CustomSurvey, CustomHousehold, CustomHome, C
     interview: UserInterviewAttributes<CustomSurvey, CustomHousehold, CustomHome, CustomPerson>;
     // TODO There's also a path in widgetConfig, but this one comes from the props of the question. See if it's always the same and use the one from widgetConfig if necessary
     path: string;
-    user: FrontendUser;
+    user: CliUser;
     inputRef?: React.LegacyRef<HTMLTextAreaElement>;
     size?: 'small' | 'medium' | 'large';
     widgetConfig: InputCheckboxType<CustomSurvey, CustomHousehold, CustomHome, CustomPerson>;
