+++ packages/evolution-frontend/src/components/inputs/InputMapFindPlace.tsx
@@ -14,10 +14,10 @@ import remarkGfm from 'remark-gfm';
 
 import { _isBlank } from 'chaire-lib-common/lib/utils/LodashExtensions';
 import projectConfig from 'chaire-lib-common/lib/config/shared/project.config';
-import { InputMapFindPlaceType } from '../../services/widgets';
-import * as surveyHelper from '../../utils/helpers';
+import { InputMapFindPlaceType } from 'evolution-common/lib/services/widgets';
+import * as surveyHelper from 'evolution-common/lib/utils/helpers';
 import { UserInterviewAttributes } from 'evolution-common/lib/services/interviews/interview';
-import { FrontendUser } from 'chaire-lib-frontend/lib/services/auth/user';
+import { CliUser } from 'chaire-lib-common/lib/services/user/userType';
 import { WithTranslation, withTranslation } from 'react-i18next';
 import { FeatureGeocodedProperties, MarkerData, PlaceGeocodedProperties } from './maps/InputMapTypes';
 import InputSelect from './InputSelect';
@@ -34,7 +34,7 @@ export interface InputMapFindPlaceProps<CustomSurvey, CustomHousehold, CustomHom
     // TODO There's also a path in widgetConfig, but this one comes from the props of the question. See if it's always the same and use the one from widgetConfig if necessary
     path: string;
     loadingState?: number;
-    user: FrontendUser;
+    user: CliUser;
     shortname?: string;
     inputRef?: React.LegacyRef<HTMLInputElement>;
     size?: 'small' | 'medium' | 'large';
