+++ packages/evolution-frontend/src/components/inputs/InputMultiselect.tsx
@@ -10,10 +10,10 @@ import Select from 'react-select';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { _isBlank } from 'chaire-lib-common/lib/utils/LodashExtensions';
-import * as surveyHelper from '../../utils/helpers';
-import { ChoiceType, InputMultiselectType } from '../../services/widgets';
+import * as surveyHelper from 'evolution-common/lib/utils/helpers';
+import { ChoiceType, InputMultiselectType } from 'evolution-common/lib/services/widgets';
 import { UserInterviewAttributes } from 'evolution-common/lib/services/interviews/interview';
-import { FrontendUser } from 'chaire-lib-frontend/lib/services/auth/user';
+import { CliUser } from 'chaire-lib-common/lib/services/user/userType';
 
 interface InputMultiselectProps<CustomSurvey, CustomHousehold, CustomHome, CustomPerson> {
     id: string;
@@ -22,7 +22,7 @@ interface InputMultiselectProps<CustomSurvey, CustomHousehold, CustomHome, Custo
     interview: UserInterviewAttributes<CustomSurvey, CustomHousehold, CustomHome, CustomPerson>;
     // TODO There's also a path in widgetConfig, but this one comes from the props of the question. See if it's always the same and use the one from widgetConfig if necessary
     path: string;
-    user: FrontendUser;
+    user: CliUser;
     inputRef?: React.LegacyRef<HTMLInputElement>;
     size?: 'small' | 'medium' | 'large';
     widgetConfig: InputMultiselectType<CustomSurvey, CustomHousehold, CustomHome, CustomPerson>;
