--- packages/evolution-frontend/src/utils/helpers/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright 2022, Polytechnique Montreal and contributors
- *
- * This file is licensed under the MIT License.
- * License text available at https://opensource.org/licenses/MIT
- */
-export * from 'evolution-common/lib/utils/helpers';
-export * from './frontendHelpers';
