+++ packages/evolution-legacy/src/components/survey/GroupedObject.js
@@ -11,7 +11,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashAlt }      from '@fortawesome/free-solid-svg-icons/faTrashAlt';
 import _get                from 'lodash.get';
 
-import * as surveyHelperNew from 'evolution-frontend/lib/utils/helpers';
+import * as surveyHelperNew from 'evolution-common/lib/utils/helpers';
 import { withSurveyContext } from 'evolution-frontend/lib/components/hoc/WithSurveyContextHoc';
 import ConfirmModal     from './modal/ConfirmModal';
 import Text             from './Text';
