+++ packages/evolution-legacy/src/components/survey/Section.js
@@ -11,7 +11,7 @@ import _get                     from 'lodash.get';
 import _shuffle                 from 'lodash.shuffle';
 
 import sectionTemplate  from './SectionTemplateHOC';
-import * as surveyHelper from 'evolution-frontend/lib/utils/helpers';
+import * as surveyHelper from 'evolution-common/lib/utils/helpers';
 import { withSurveyContext } from 'evolution-frontend/lib/components/hoc/WithSurveyContextHoc';
 import Text             from './Text';
 import InfoMap          from './InfoMap';
